+++ BrowserPane.m
@@ -249,7 +249,7 @@
  * Called from the webview when the user positions the mouse over an element. If it's a link
  * then echo the URL to the status bar like Safari does.
  */
--(void)webView:(WebView *)sender mouseDidMoveOverElement:(NSDictionary *)elementInformation modifierFlags:(unsigned int)modifierFlags
+-(void)webView:(WebView *)sender mouseDidMoveOverElement:(NSDictionary *)elementInformation modifierFlags:(NSUInteger )modifierFlags
 {
 	NSURL * url = [elementInformation valueForKey:@"WebElementLinkURL"];
 	[controller setStatusMessage:(url ? [url absoluteString] : @"") persist:NO];
@@ -684,7 +684,7 @@
  * Support special key codes. If we handle the key, return YES otherwise
  * return NO to allow the framework to pass it on for default processing.
  */
--(BOOL)handleKeyDown:(unichar)keyChar withFlags:(unsigned int)flags
+-(BOOL)handleKeyDown:(unichar)keyChar withFlags:(NSUInteger )flags
 {
 	return NO;
 }
